+++ Timing/Presentation/Controls/TimingGridControl/TimingGrid.xaml
@@ -102,6 +102,7 @@
         <Setter Property="Foreground" Value="{StaticResource LightGrey01Brush}"></Setter>
         <Setter Property="ColumnHeaderStyle" Value="{StaticResource DataGridHeaderStyle}"/>
         <Setter Property="HeadersVisibility" Value="Column"/>
+        <Setter Property="BorderBrush" Value="Transparent"/>
         <Setter Property="IsReadOnly" Value="True"/>
         <Setter Property="SelectionMode" Value="Single"/>
         <Setter Property="SelectionUnit" Value="FullRow"/>
