+++ WindowsControls/WPF/Constants.xaml
@@ -5,6 +5,7 @@
     <system:Double x:Key="WheelControlCoreTempFontSize">25</system:Double>
     <system:Double x:Key="GearIndicationFontSize">25</system:Double>
     <system:Double x:Key="BrakeTemperatureFontSize">23</system:Double>
+    <system:Double x:Key="ExtraBigFontSize">28</system:Double>
     <system:Double x:Key="BigFontSize">20</system:Double>
     <system:Double x:Key="NormalFontSize">15</system:Double>
     <system:Double x:Key="TimingGridFontSize">18</system:Double>
